[Firtool] Move to lib/Tools, NFC
