+++ lib/CMakeLists.txt
@@ -7,7 +7,6 @@ add_subdirectory(Bindings)
 add_subdirectory(CAPI)
 add_subdirectory(Conversion)
 add_subdirectory(Dialect)
-add_subdirectory(Firtool)
 add_subdirectory(Reduce)
 add_subdirectory(Scheduling)
 add_subdirectory(Support)
